+++ src/pages/VulnerabilityScan.jsx
@@ -6,6 +6,7 @@ import "react-toastify/dist/ReactToastify.css";
 //import { useNavigate } from "react-router-dom";
 import { Line } from "react-chartjs-2";
 import { analyzeScanData } from "../utils/groq";
+import { useScan } from "../context/ScanContext";
 
 import {
   Chart as ChartJS,
@@ -40,14 +41,20 @@ const API_KEY = import.meta.env.VITE_ZAP_API_KEY;
 const BASE_URL = "/zap";
 
 function VulnerabilityScan() {
-  const [pluginProgressData, setPluginProgressData] = useState([]);
+const {
+  targetUrl,
+  setTargetUrl,
+  scanProgress,
+  alerts,
+  sentMessages,
+  pluginProgressData,
+  scanStatus,
+  loading,
+  startScan,
+  stopScan,
+} = useScan();
 
-  const [targetUrl, setTargetUrl] = useState("");
-  const [scanStatus, setScanStatus] = useState("");
-  const [alerts, setAlerts] = useState([]);
-  const [sentMessages, setSentMessages] = useState([]);
-  const [loading, setLoading] = useState(false);
-  const [scanProgress, setScanProgress] = useState(0);
+
   const pollingRef = useRef(true);
   //const navigate = useNavigate();
   const [showProgress, setShowProgress] = useState(false);
@@ -56,280 +63,6 @@ function VulnerabilityScan() {
   const [showAIReport, setShowAIReport] = useState(false);
   const [aiReport, setAIReport] = useState("");
 
-  const isValidUrl = (url) => {
-    try {
-      new URL(url);
-      return true;
-    } catch {
-      return false;
-    }
-  };
-
-  const startScan = async () => {
-    if (!isValidUrl(targetUrl)) {
-      setScanStatus("Please enter a valid URL including http/https.");
-      toast("Invalid URL. Please include http/https.", {
-        type: "error",
-        theme: "colored",
-        autoClose: 5000,
-        transition: Slide,
-      });
-      return;
-    }
-
-    toast(`Started scan on ${targetUrl}`, {
-      type: "info",
-      theme: "colored",
-      autoClose: 5000,
-      transition: Slide,
-    });
-
-    setLoading(true);
-    setAlerts([]);
-    setSentMessages([]);
-    setScanProgress(0);
-    pollingRef.current = true;
-
-    try {
-      const spiderResp = await fetch(
-        `${BASE_URL}/spider/action/scan/?url=${encodeURIComponent(
-          targetUrl
-        )}&apikey=${API_KEY}`
-      );
-      const spiderData = await spiderResp.json();
-      await pollSpiderStatus(spiderData.scan);
-
-      const activeResp = await fetch(
-        `${BASE_URL}/ascan/action/scan/?url=${encodeURIComponent(
-          targetUrl
-        )}&apikey=${API_KEY}`
-      );
-      const activeData = await activeResp.json();
-
-      if (activeData.scan) {
-        await pollScanStatus(activeData.scan);
-      }
-
-      toast(`Scan completed for ${targetUrl}`, {
-        type: "success",
-        theme: "colored",
-        autoClose: 5000,
-        transition: Slide,
-      });
-    } catch (error) {
-      setScanStatus(`Error starting scan: ${error.message}`);
-      toast(`Scan failed: ${error.message}`, {
-        type: "error",
-        theme: "colored",
-        autoClose: 5000,
-        transition: Slide,
-      });
-    }
-
-    setLoading(false);
-  };
-
-  const pollSpiderStatus = async (scanId) => {
-    let status = "0";
-    while (status !== "100" && pollingRef.current) {
-      await new Promise((res) => setTimeout(res, 500));
-      const resp = await fetch(
-        `${BASE_URL}/spider/view/status/?scanId=${scanId}&apikey=${API_KEY}`
-      );
-      status = (await resp.json()).status;
-    }
-  };
-
-  const pollScanStatus = async (scanId) => {
-    let status = "0";
-
-    while (status !== "100" && pollingRef.current) {
-      await new Promise((res) => setTimeout(res, 1000)); // 1-second interval
-
-      // Fetch scan progress
-      const resp = await fetch(
-        `${BASE_URL}/ascan/view/status/?scanId=${scanId}&apikey=${API_KEY}`
-      );
-      status = (await resp.json()).status;
-      setScanProgress(Number(status));
-
-      // Fetch plugin scan progress in real-time
-      const plugins = await fetchPluginProgress();
-      setPluginProgressData(plugins);
-
-      // Fetch HTTP messages
-      const msgs = await fetchSentMessages();
-      setSentMessages(msgs);
-    }
-
-    // Final fetch after scan completes
-    if (pollingRef.current) {
-      setScanProgress(100);
-      await fetchAlerts();
-      await fetchSentMessages();
-    }
-  };
-
-  const stopScan = async () => {
-    pollingRef.current = false; // Stop further polling
-    setLoading(false); // Hide the stop button
-    setScanStatus("Scan stopped by user.");
-
-    try {
-      // Stop scans on ZAP
-      await fetch(`${BASE_URL}/ascan/action/stop/?apikey=${API_KEY}`);
-      await fetch(`${BASE_URL}/spider/action/stop/?apikey=${API_KEY}`);
-
-      // Fetch the latest results
-      await fetchAlerts();
-      const msgs = await fetchSentMessages();
-      setSentMessages(msgs);
-
-      toast("Scan stopped and results fetched.", {
-        type: "warning",
-        theme: "colored",
-        autoClose: 4000,
-        transition: Slide,
-      });
-    } catch (error) {
-      toast(`Error stopping scan: ${error.message}`, {
-        type: "error",
-        theme: "colored",
-        autoClose: 4000,
-        transition: Slide,
-      });
-    }
-  };
-
-  const fetchAlerts = async () => {
-    try {
-      const resp = await fetch(
-        `${BASE_URL}/core/view/alerts/?baseurl=${encodeURIComponent(
-          targetUrl
-        )}&apikey=${API_KEY}`
-      );
-      const data = await resp.json();
-      setAlerts(data.alerts || []);
-      setScanStatus(`Found ${data.alerts?.length || 0} alerts.`);
-      toast.success(`Found ${data.alerts?.length || 0} vulnerabilities`, {
-        position: "bottom-center",
-        autoClose: 2500,
-      });
-      return data.alerts || []; // ✅ return alerts
-    } catch (error) {
-      setScanStatus(`Error fetching alerts: ${error.message}`);
-      toast.error(`Error fetching alerts: ${error.message}`, {
-        position: "top-center",
-        autoClose: 2500,
-      });
-      return [];
-    }
-  };
-
-  const fetchPluginProgress = async () => {
-    try {
-      const response = await fetch(
-        `${BASE_URL}/ascan/view/scanners/?apikey=${API_KEY}`
-      );
-      const json = await response.json();
-      const scanners = json.scanners || [];
-
-      console.log("Scanners fetched:", scanners); // <-- ADD THIS
-
-      return scanners.map((scanner) => ({
-        id: scanner.id,
-        name: scanner.name,
-        strength: scanner.attackStrength || "-",
-        progress: parseInt(scanner.progress),
-        elapsed: scanner.timeStarted
-          ? getElapsedTime(scanner.timeStarted)
-          : "-",
-        requests: scanner.requests || 0,
-        alerts: scanner.alertsRaised || 0,
-        status: scanner.progress >= 100 ? "✅ Done" : "⏳ Running",
-      }));
-    } catch (error) {
-      console.error("Failed to fetch plugin scanners", error);
-      return [];
-    }
-  };
-
-  // Utility function to format elapsed time
-  function getElapsedTime(startTime) {
-    const start = new Date(startTime);
-    const now = new Date();
-    const diff = now - start;
-    const seconds = (diff / 1000).toFixed(3);
-    const minutes = Math.floor(seconds / 60);
-    const remSeconds = (seconds % 60).toFixed(3).padStart(6, "0");
-    return `00:${String(minutes).padStart(2, "0")}:${remSeconds}`;
-  }
-
-  const fetchSentMessages = async () => {
-    try {
-      const resp = await fetch(
-        `${BASE_URL}/core/view/messages/?baseurl=${encodeURIComponent(
-          targetUrl
-        )}&apikey=${API_KEY}`
-      );
-
-      if (!resp.ok) {
-        throw new Error(`ZAP responded with status: ${resp.status}`);
-      }
-
-      const data = await resp.json();
-
-      const messages = data.messages || [];
-      if (messages.length === 0) {
-        console.warn("No messages found for target:", targetUrl);
-        return [];
-      }
-
-      const detailedMessages = await Promise.all(
-        messages.map(async (msg) => {
-          try {
-            const msgResp = await fetch(
-              `${BASE_URL}/core/view/message/?id=${msg.id}&apikey=${API_KEY}`
-            );
-
-            if (!msgResp.ok) {
-              throw new Error(
-                `Failed to fetch message ${msg.id}: ${msgResp.status}`
-              );
-            }
-
-            const msgData = await msgResp.json();
-            return {
-              id: msg.id,
-              timeSent: msgData.message.timeSent,
-              timeReceived: msgData.message.timeReceived,
-              method: msgData.message.requestHeader?.split(" ")[0] || "-",
-              url: msgData.message.requestHeader?.split(" ")[1] || "-",
-              statusCode:
-                msgData.message.responseHeader?.match(
-                  /HTTP\/\d\.\d (\d{3})/
-                )?.[1] || "-",
-              reasonPhrase:
-                msgData.message.responseHeader?.match(
-                  /HTTP\/\d\.\d \d{3} (.+)/
-                )?.[1] || "-",
-              rtt: msgData.message.rtt,
-              responseHeader: msgData.message.responseHeader,
-              responseBody: msgData.message.responseBody,
-            };
-          } catch (err) {
-            console.error(`Error fetching message ${msg.id}:`, err);
-            return { id: msg.id };
-          }
-        })
-      );
-
-      return detailedMessages;
-    } catch (error) {
-      console.error("Failed to fetch sent messages from ZAP:", error.message);
-      return [];
-    }
-  };
 
   const getRiskCounts = () => {
     const counts = {
